+++ tests/tests.rs
@@ -18,7 +18,7 @@ use tokio::time::sleep;
 
 mod common;
 
-use common::KeyTest;
+use common::{KeyTest, MAX_DEFER_TIME, MIN_DEFER_TIME};
 
 #[test]
 fn test_hash_algorithm_compat() {
@@ -893,8 +893,8 @@ async fn test_mark_as_deleted_single() {
     let path = common::init("mark_as_deleted_single");
     let storage = common::create_test_storage(&path, 10_000).await.unwrap();
     let count = 30;
-    let delete_key = KeyTest::new(5);
     let records = common::generate_records(count, 1000);
+    let delete_key = KeyTest::new(records[0].0);
     for (key, data) in &records {
         write_one(&storage, *key, data, None).await.unwrap();
         sleep(Duration::from_millis(64)).await;
@@ -903,6 +903,38 @@ async fn test_mark_as_deleted_single() {
     assert!(!storage.contains(delete_key).await.unwrap());
     common::clean(storage, path).await.expect("clean failed");
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
+}
+
+#[tokio::test]
+async fn test_mark_as_deleted_deferred_dump() {
+    let path = common::init("mark_as_deleted_deferred_dump");
+    let storage = common::create_test_storage(&path, 10_000).await.unwrap();
+    let count = 30;
+    let records = common::generate_records(count, 1000);
+    let delete_key = KeyTest::new(records[0].0);
+    for (key, data) in &records {
+        write_one(&storage, *key, data, None).await.unwrap();
+        sleep(Duration::from_millis(64)).await;
+    }
+    let _ = storage.close().await;
+
+    let storage = common::create_test_storage(&path, 10_000).await.unwrap();
+    let update_time = std::fs::metadata(&path.join("test.0.index")).expect("metadata");
+    storage.mark_all_as_deleted(&delete_key).await.unwrap();
+
+    sleep(MIN_DEFER_TIME / 2).await;
+    let new_update_time = std::fs::metadata(&path.join("test.0.index")).expect("metadata");
+    assert_eq!(
+        update_time.modified().unwrap(),
+        new_update_time.modified().unwrap()
+    );
+    sleep(MAX_DEFER_TIME).await;
+    let new_update_time = std::fs::metadata(&path.join("test.0.index")).expect("metadata");
+    assert_ne!(
+        update_time.modified().unwrap(),
+        new_update_time.modified().unwrap()
+    );
+    common::clean(storage, path).await.expect("clean failed");
 }
 
 #[tokio::test]