+++ CHANGELOG.md
@@ -7,9 +7,11 @@ Pearl changelog
 - Add method to receive occupied disk space (#170)
 
 #### Changed
+- Defer index drop on delete (#165)
 
 #### Fixed
 - Restore documentation publishing (#173)
+
 
 #### Updated